requirements.txt
django
djangorestframework
django-cors-headers
backports.zoneinfo
psycopg2
gradio
transformers
fuzzywuzzy
readability
python-Levenshtein
pandas
nltk
numpy
torch
gensim
scikit-learn
seaborn
matplotlib